+++ src/main.tsx
@@ -1,9 +1,9 @@
-import '@/styles/index.css';
+import '@/index.css';
 
 import React from 'react';
 import ReactDOM from 'react-dom/client';
 
-import App from '@/App';
+import App from '@/components/App';
 import { AuthProvider } from '@/contexts/AuthContexts';
 
 // eslint-disable-next-line @typescript-eslint/no-non-null-assertion
